chore(Qualtrics): move sdk loader to different file
